pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi[all]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.40",
]
